requirements.fixed.txt
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.0
cryptography==41.0.7
Deprecated==1.2.14
exceptiongroup==1.2.0
Flask==3.0.0
github-app-handler==0.9.1
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
requests==2.31.0
sentry-sdk==1.39.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
